+++ src/__tests__/modules/document.test.ts
@@ -9,6 +9,7 @@ import Document from '../../model/document'
 import { computeLinesOffsets, LinesTextDocument } from '../../model/textdocument'
 import { disposeAll } from '../../util'
 import workspace from '../../workspace'
+import events from '../../events'
 import helper from '../helper'
 
 let nvim: Neovim
@@ -233,6 +234,7 @@ describe('Document', () => {
     })
 
     it('should move cursor', async () => {
+      await nvim.call('cursor', [1, 1])
       let doc = await workspace.document
       let edits: TextEdit[] = []
       edits.push({
@@ -240,6 +242,9 @@ describe('Document', () => {
         newText: 'foo'
       })
       await doc.applyEdits(edits, false, true)
+      let cursor = await nvim.call('getcurpos') as number[]
+      expect(cursor[1]).toBe(1)
+      expect(cursor[2]).toBe(3)
     })
 
     it('should applyEdits with range not sorted', async () => {